Moved twitter client logic into HomeTimelineFragment from main activity
