+++ app/src/main/java/com/codepath/apps/simpletwitterclient/activities/TimelineActivity.java
@@ -8,27 +8,12 @@ import android.view.MenuItem;
 
 import com.codepath.apps.simpletwitterclient.R;
 import com.codepath.apps.simpletwitterclient.fragments.TweetsListFragment;
-import com.codepath.apps.simpletwitterclient.lib.Logger;
-import com.codepath.apps.simpletwitterclient.lib.Network;
-import com.codepath.apps.simpletwitterclient.lib.Toaster;
-import com.codepath.apps.simpletwitterclient.models.SignedInUser;
-import com.codepath.apps.simpletwitterclient.models.Tweet;
-import com.codepath.apps.simpletwitterclient.models.User;
-import com.codepath.apps.simpletwitterclient.networking.TwitterApplication;
-import com.codepath.apps.simpletwitterclient.networking.TwitterClient;
-import com.loopj.android.http.JsonHttpResponseHandler;
-
-import org.apache.http.Header;
-import org.json.JSONArray;
-import org.json.JSONObject;
-
-import java.util.ArrayList;
 
 public class TimelineActivity extends ActionBarActivity {
 
     private final static String TAG = "TimelineActivity";
-    private TwitterClient client;
-    private long minTweetId; //Long.MAX_VALUE;
+
+
     private final int REQUEST_CODE_COMPOSE = 323;
     private TweetsListFragment fragmentTweetsList;
 
@@ -40,37 +25,22 @@ public class TimelineActivity extends ActionBarActivity {
         // Set up infinite scroll
         //setScrollListener();
 
-        //get the client
-        client = TwitterApplication.getRestClient(); //singleton client
 
 
         // Access the fragment (from layout)
-        if (savedInstanceState == null) {
-            fragmentTweetsList = (TweetsListFragment) getSupportFragmentManager().findFragmentById(R.id.fragment_timeline);
-        }
+//        if (savedInstanceState == null) {
+//            fragmentTweetsList = (TweetsListFragment) getSupportFragmentManager().findFragmentById(R.id.fragment_timeline);
+//        }
 
-        // Clear Tweets and reset state
-        clearTweets();
+
 
 
         // Pull to Refresh
         //setUpPullToRefresh();
 
-        loadNewTweets();
     }
 
-    /**
-     *  Clear current Tweets
-     */
-    private void clearTweets() {
-        // Tried to use Long.MAX_VALUE here, but didn't work. Instead used Long.MAX_VALUE/10
-        // This should be large enough
-        // Clear Current tweets
-        //aTweets.clear();
-        fragmentTweetsList.clear();
 
-        minTweetId = Long.parseLong("922337203685477580");
-    }
 
 
     @Override
@@ -91,78 +61,11 @@ public class TimelineActivity extends ActionBarActivity {
         return super.onOptionsItemSelected(item);
     }
 
-    /**
-     * Send API request to get the timeline json
-     * Fill the ListView by creating the tweet objects from json
-     */
-    private void fetchTweetsIntoTimeline(long maxTweetId) {
-
-        client.getHomeTimeline(maxTweetId, new JsonHttpResponseHandler() {
-
-            //Success
-            @Override
-            public void onSuccess(int statusCode, Header[] headers, JSONArray json) {
-
-                Logger.log(TAG, "-->network success");
-
-                // deserialize json
-                // create models and add then to the adapter
-                // load model data into listview
-                ArrayList<Tweet> tweets = Tweet.fromJsonArray(json);
-                updateMinTweetIdFromTweetList(tweets);
-                persistTweets(tweets);
-                //aTweets.addAll(tweets);
-                fragmentTweetsList.addAll(tweets);
 
 
-                //swipeContainer.setRefreshing(false);
-            }
 
-            //Failure
-            @Override
-            public void onFailure(int statusCode, Header[] headers, Throwable throwable, JSONObject errorResponse) {
 
-                try {
-                    Logger.log(TAG, "No network "+errorResponse.toString());
-                } catch (Exception e) {
-                    //do nothing
-                }
 
-                Toaster.create(TimelineActivity.this, "Sorry, the network appears to be down");
-                Toaster.create(TimelineActivity.this, "Pull to refresh to try again");
-                //swipeContainer.setRefreshing(false);
-            }
-        });
-    }
-
-    /**
-     * Keeps track of the minimum tweet id so that new non-dupe tweets can be fetched
-     */
-    private void updateMinTweetIdFromTweetList(ArrayList<Tweet> tweetsArray) {
-        long curTweetId;
-
-        for(int i = 0; i < tweetsArray.size() ; i++) {
-            curTweetId = tweetsArray.get(i).getUid();
-            if (curTweetId < minTweetId) {
-                minTweetId = curTweetId;
-            }
-        }
-    }
-
-    /**
-     * Persists each Tweet into SQLite via Active Android
-     */
-    private void persistTweets(ArrayList<Tweet> tweetsArray) {
-        for(int i = 0; i < tweetsArray.size() ; i++) {
-            Tweet curTweet = tweetsArray.get(i);
-
-            // Save Tweet's user
-            curTweet.getUser().save();
-
-            // Save Tweet
-            curTweet.save();
-        }
-    }
 
     /**
      * Sets up infinite scrolling
@@ -194,61 +97,18 @@ public class TimelineActivity extends ActionBarActivity {
 
             // If Tweeted, then clear the current tweets and re-fetch new ones
             if (isSuccess) {
-                clearTweets();
-                fetchTweetsIntoTimeline(minTweetId);
+                // need to call these on the fragment?
+                //clearTweets();
+                //fetchTweetsIntoTimeline(minTweetId);
             }
         }
     }
 
-    /**
-     * This fetch's the signed in user's profile
-     */
-    private void fetchSignedInUsersProfile() {
 
-        client.getUserProfile(new JsonHttpResponseHandler() {
 
-            //Success
-            @Override
-            public void onSuccess(int statusCode, Header[] headers, JSONObject json) {
-                User signedInUser = User.fromJson(json);
-                SignedInUser.persistSignedInUser(TimelineActivity.this, signedInUser);
-            }
 
-            //Failure
-            @Override
-            public void onFailure(int statusCode, Header[] headers, Throwable throwable, JSONObject errorResponse) {
-                // Do Nothing
-            }
-        });
-    }
 
-    /**
-     * This loads new tweets for the first time
-     */
-    private void loadNewTweets() {
 
-        Network network = new Network();
-        if (network.isNetworkAvailable(this)) {
-            fetchTweetsIntoTimeline(minTweetId);
-            fetchSignedInUsersProfile();
-
-        } else {
-            Toaster.create(TimelineActivity.this, "Sorry, the network appears to be down. Showing cached data");
-            Toaster.create(TimelineActivity.this, "Pull to refresh to try again");
-            loadTweetsFromCache();
-            //swipeContainer.setRefreshing(false);
-        }
-    }
-
-    /**
-     * Loads tweets from SQLite cache. This is used in the case of no network
-     */
-    private void loadTweetsFromCache() {
-        // Get all tweets from Storage
-        ArrayList existingTweets = (ArrayList) Tweet.getAll();
-        //aTweets.addAll(existingTweets);
-        fragmentTweetsList.addAll(existingTweets);
-    }
 
     /**
      * Sets up the pull to refresh functionality
